refactor: InputStepper Storybook 동작 구현
